Require the ActiveSupport extensions needed
Rather than requiring the extensions in the tests, we can just require what we
need in the place we need it.

New tests aren't needed - the original tests were passing because necessary
dependencies were loaded in the test_helper rather than in the runtime code.

ActiveSupport is now a runtime dependency as it's now required by runtime
files, not test files.